Abstract out useModelStream hook
